% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RobustSE_function_ZY_incomplete.R
\name{vcovCR.glmerModZY}
\alias{vcovCR.glmerModZY}
\title{Variance Covariance Computation for GLMM}
\arguments{
\item{obj}{A \code{glmerMod} object representing the fitted model.}

\item{cluster}{Optional; a vector indicating the cluster structure within the data.
If NULL, attempts to retrieve cluster information from attributes of the model object.}

\item{type}{Character string specifying the type of variance-covariance matrix to compute.
Supported types include "classic", with potential for further types depending on method implementation.}
}
\value{
Returns a diagonal matrix with the computed variances for each fixed effect in the model.
}
\description{
Computes robust variance-covariance matrix for generalized linear mixed models
fitted with \code{glmer} from the \code{lme4} package. It allows specification of
clustering and supports different types of variance estimations.
}
